+++ spdmlib/src/crypto/spdm_ring/cert_operation_impl.rs
@@ -6,10 +6,58 @@ extern crate alloc;
 
 use crate::crypto::SpdmCertOperation;
 use crate::error::{SpdmResult, SPDM_STATUS_INVALID_CERT};
+use ring::signature::{self, UnparsedPublicKey};
+use spdm_x509::crypto_backend::{CryptoBackend, SignatureAlgorithm};
+use spdm_x509::error::{Error as X509Error, Result as X509Result, SignatureError};
 
 // =============================================================================
-// spdm_x509 implementation
+// ring CryptoBackend for spdm_x509 certificate-chain validation.
+//
+// spdm_x509 ships no built-in backend; each crypto crate supplies its own. This
+// backend verifies classical (RSA/ECDSA/Ed25519) certificate signatures via
+// ring. Post-quantum (ML-DSA / FIPS 204) signatures are dispatched by
+// spdm_x509's validator to the registered PQC verifier hook (e.g. from
+// spdmlib_crypto_aws_lc) and never reach this backend.
 // =============================================================================
+
+/// ring-based crypto backend for X.509 signature verification.
+#[derive(Clone, Copy)]
+pub struct RingBackend;
+
+impl CryptoBackend for RingBackend {
+    fn verify_signature(
+        &self,
+        algorithm: SignatureAlgorithm,
+        tbs_data: &[u8],
+        signature: &[u8],
+        public_key: &[u8],
+    ) -> X509Result<()> {
+        let ring_algo: &dyn signature::VerificationAlgorithm = match algorithm {
+            SignatureAlgorithm::EcdsaP256Sha256 => &signature::ECDSA_P256_SHA256_ASN1,
+            SignatureAlgorithm::EcdsaP256Sha384 => &signature::ECDSA_P256_SHA384_ASN1,
+            SignatureAlgorithm::EcdsaP384Sha256 => &signature::ECDSA_P384_SHA256_ASN1,
+            SignatureAlgorithm::EcdsaP384Sha384 => &signature::ECDSA_P384_SHA384_ASN1,
+            SignatureAlgorithm::RsaPkcs1Sha256 => &signature::RSA_PKCS1_2048_8192_SHA256,
+            SignatureAlgorithm::RsaPkcs1Sha384 => &signature::RSA_PKCS1_2048_8192_SHA384,
+            SignatureAlgorithm::RsaPkcs1Sha512 => &signature::RSA_PKCS1_2048_8192_SHA512,
+            SignatureAlgorithm::RsaPssSha256 => &signature::RSA_PSS_2048_8192_SHA256,
+            SignatureAlgorithm::RsaPssSha384 => &signature::RSA_PSS_2048_8192_SHA384,
+            SignatureAlgorithm::RsaPssSha512 => &signature::RSA_PSS_2048_8192_SHA512,
+            SignatureAlgorithm::Ed25519 => &signature::ED25519,
+            SignatureAlgorithm::MlDsa44
+            | SignatureAlgorithm::MlDsa65
+            | SignatureAlgorithm::MlDsa87 => {
+                return Err(X509Error::unsupported_algorithm(
+                    "ML-DSA is not supported by the ring backend",
+                ));
+            }
+        };
+
+        let pk = UnparsedPublicKey::new(ring_algo, public_key);
+        pk.verify(tbs_data, signature)
+            .map_err(|_| X509Error::SignatureError(SignatureError::VerificationFailed))
+    }
+}
 
 pub static DEFAULT: SpdmCertOperation = SpdmCertOperation {
     get_cert_from_cert_chain_cb: get_cert_from_cert_chain_x509,
@@ -26,8 +74,9 @@ fn verify_cert_chain_x509(
     base_asym_algo: Option<u32>,
     base_hash_algo: Option<u32>,
 ) -> SpdmResult {
-    spdm_x509::x509::chain::verify_cert_chain_with_options(
+    spdm_x509::x509::chain::verify_cert_chain_with_backend(
         cert_chain,
+        RingBackend,
         base_asym_algo,
         base_hash_algo,
     )
