+++ spdmlib/src/crypto/spdm_x509/Cargo.toml
@@ -15,11 +15,6 @@ der = { version = "0.7", default-features = false, features = ["alloc", "derive"
 spki = { version = "0.7", default-features = false, features = ["alloc"] }
 const-oid = { version = "0.9", default-features = false }
 
-# Cryptographic backends (optional)
-ring = { version = "0.17", optional = true, default-features = false, features = ["alloc"] }
-untrusted = { version = "0.9", optional = true }
-mbedtls = { version = "0.9.1", optional = true, default-features = false, features = ["no_std_deps", "rdrand"] }
-
 # PEM encoding support
 pem-rfc7468 = { version = "0.7", default-features = false, features = ["alloc"] }
 
@@ -37,12 +32,16 @@ log = { version = "0.4", default-features = false }
 hex = "0.4"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
+# Test-only crypto backend: `ring` powers the bundled RingBackend, which is gated
+# on `cfg(test)`, so it exists only in this crate's own test builds and can never
+# be linked into a production build. Production consumers supply their own
+# `CryptoBackend`.
+ring = { version = "0.17", default-features = false, features = ["alloc", "std"] }
+untrusted = "0.9"
 
 [features]
-default = ["ring-backend"]
-ring-backend = ["ring", "untrusted"]
-mbedtls-backend = ["mbedtls"]
-std = ["der/std", "ring?/std", "log/std", "conquer-once/std"]
+default = []
+std = ["der/std", "log/std", "conquer-once/std"]
 alloc = ["der/alloc"]
 
 [lib]
