+++ spdmlib_crypto_aws_lc/Cargo.toml
@@ -24,9 +24,9 @@ spin = "0.9.8"
 bytes = { version = "1", default-features = false }
 
 [dev-dependencies]
-# Enable the ring backend + std so the cert-chain tests can read test-key DER
-# files and exercise the full validator.
-spdm_x509 = { path = "../spdmlib/src/crypto/spdm_x509", default-features = false, features = ["ring-backend", "std"] }
+# std so the cert-chain tests can read test-key DER files. No ring-backend: the
+# tests drive validation through AwsLcBackend (classical + ML-DSA).
+spdm_x509 = { path = "../spdmlib/src/crypto/spdm_x509", default-features = false, features = ["std"] }
 
 [features]
 default = []
