+++ spdmlib_crypto_mbedtls/Cargo.toml
@@ -7,16 +7,23 @@ license = "Apache-2.0 or MIT"
 [dependencies]
 spdmlib = { path = "../spdmlib", default-features = false}
 # spdm_x509 provides ML-DSA (FIPS 204) aware certificate-chain validation via
-# its pluggable crypto backend plus the runtime PQC verifier hook. Use the
-# mbedtls backend so classical (RSA/ECC) certificate signatures are verified by
-# mbedtls and no ring is linked; post-quantum signatures still dispatch to the
+# its pluggable crypto backend plus the runtime PQC verifier hook. spdm_x509
+# ships no built-in backend, so this crate supplies its own `MbedtlsBackend`
+# (src/x509_backend.rs): classical (RSA/ECC) certificate signatures are verified
+# by mbedtls and no ring is linked; post-quantum signatures still dispatch to the
 # registered PQC verifier hook, so PQC certificate chains keep working.
-spdm_x509 = { path = "../spdmlib/src/crypto/spdm_x509", default-features = false, features = ["mbedtls-backend"] }
+spdm_x509 = { path = "../spdmlib/src/crypto/spdm_x509", default-features = false }
 lazy_static = { version = "1.0", features = ["spin_no_std"] }
 spin = "0.9.8"
 zeroize = { version = "1.5.0", features = ["zeroize_derive"]}
 mbedtls = { version = "0.9.1", default-features = false, features = ["no_std_deps", "rdrand"]}
 mbedtls-sys-auto = { version = "2.28.0", default-features = false }
+# Used by the mbedtls X.509 backend to rebuild a SubjectPublicKeyInfo DER around
+# the raw issuer public key so mbedtls can parse it.
+der = { version = "0.7", default-features = false, features = ["alloc", "oid"] }
+spki = { version = "0.7", default-features = false, features = ["alloc"] }
+const-oid = { version = "0.9", default-features = false }
+log = { version = "0.4", default-features = false }
 
 [features]
 default = ["hashed-transcript-data", "std"]
